refactored edit workin
