+++ components/Header/Header.js
@@ -7,7 +7,7 @@ export default function Header() {
   header.innerHTML = /* html */ `
 
   <img class="logo" src="images/timetologo600neu.png" alt="Logo time to work or do what you want" />
-  <span class="version">v0.058</span>
+  <span class="version">v0.065</span>
   `;
   header.append(Menu());
   return header;
