+++ utils.js
@@ -1,19 +1,17 @@
-const dateOptions = {
-  weekday: "long",
-  year: "numeric",
-  month: "long",
-  day: "numeric",
-};
+import Modal from "./components/Modal/Modal.js";
 
-const timeOptions = {
-  hour12: false,
-  hour: "numeric",
-  minute: "numeric",
-  second: "numeric",
-};
-
+//real utils
 function removeDuplicates(array) {
   return [...new Set(array)];
+}
+
+function generateRandomInteger(max, min) {
+  const minNumber = Number(min) || 0;
+  const maxNumber = Number(max);
+  const range = Math.abs(Math.round(maxNumber - minNumber));
+  const withinRange = Math.round(Math.random() * range);
+  const randomInteger = minNumber + withinRange;
+  return randomInteger;
 }
 
 function saveToLocalStorage(data, name) {
@@ -21,20 +19,70 @@ function saveToLocalStorage(data, name) {
   localStorage.setItem(backupName, JSON.stringify(data));
 }
 
-function formReset() {
-  const form = document.getElementById("main-form");
-  form.reset();
+//list utils (may move these)
+function loadBackup() {
+  const userEntries = JSON.parse(localStorage.getItem("TasksBackup"));
+  const timeRecords = document.getElementById("time-records");
+  timeRecords.replaceWith(timeRecords(userEntries));
+  saveToLocalStorage(userEntries, "RecordedTasks");
 }
 
 function deleteEntry(id) {
-  const userEntries = JSON.parse(localStorage.getItem("RecordedTasks"));
-  const updatedEntries = userEntries.filter((entry) => entry.id != id);
+  const recordedTasks = JSON.parse(localStorage.getItem("RecordedTasks"));
+  const updatedEntries = recordedTasks.filter((entry) => entry.id != id);
   const entryToDelete = document.getElementById(`${id}`);
   entryToDelete.remove();
   localStorage.setItem("RecordedTasks", JSON.stringify(updatedEntries));
   const listContainer = document.getElementById("list-container");
   updatedEntries.length === 0 &&
     listContainer.appendChild(Article(noTasksContent));
+}
+
+function deleteAllEntries() {
+  const updatedEntries = [];
+  localStorage.setItem("RecordedTasks", JSON.stringify(updatedEntries));
+  const modal = document.getElementById("modal1");
+  modal.remove();
+  const timeRecords = document.getElementById("time-records");
+  timeRecords.replaceWith(Article(noTasksContent));
+}
+
+function handleEdit(editId) {
+  const editForm = document.getElementById("edit-form");
+  const formData = new FormData(editForm);
+  const data = Object.fromEntries(formData);
+
+  const currentDate = new Date();
+
+  const recordedTasks = JSON.parse(localStorage.getItem("RecordedTasks"));
+  const entryToEdit = recordedTasks.filter((entry) => entry.id === editId)[0];
+
+  //needs converting from timespent to timespan(ms) but not strictly necessary now
+  const changedEntry = {
+    ...entryToEdit,
+    project: data.project,
+    task: data.task,
+    category: data.category,
+    note: data.note,
+    changeDate: currentDate,
+    timeSpent: data.timeSpent,
+  };
+
+  const editedTasks = recordedTasks.map((entry) =>
+    entry.id === editId ? changedEntry : entry
+  );
+
+  localStorage.setItem("RecordedTasks", JSON.stringify(editedTasks));
+  const edit = document.getElementById("edit-container");
+  edit.remove();
+  const app = document.getElementById("app");
+  app.append(Modal("afterEdit"));
+}
+
+//rest (form)
+function formReset() {
+  const form = document.getElementById("main-form");
+  form.reset();
 }
 
 function timeReset() {
@@ -49,14 +97,33 @@ function timeReset() {
   sector1.classList.add("sector--passive");
 }
 
-function generateRandomInteger(max, min) {
-  const minNumber = Number(min) || 0;
-  const maxNumber = Number(max);
-  const range = Math.abs(Math.round(maxNumber - minNumber));
-  const withinRange = Math.round(Math.random() * range);
-  const randomInteger = minNumber + withinRange;
-  return randomInteger;
+function createUnixTimeID(unixTime) {
+  const unixDate = unixTime || Date.now();
+  const firstNumber = Math.round(unixDate / 1000);
+  const firstPart = firstNumber.toString().slice(3);
+  const randomNr = Math.floor(Math.random() * 1000);
+  const secondPart = randomNr.toLocaleString("en-US", {
+    minimumIntegerDigits: 3,
+    useGrouping: false,
+  });
+  const newId = Number(firstPart + secondPart);
+  return newId;
 }
+
+// date and time
+const dateOptions = {
+  weekday: "long",
+  year: "numeric",
+  month: "long",
+  day: "numeric",
+};
+
+const timeOptions = {
+  hour12: false,
+  hour: "numeric",
+  minute: "numeric",
+  second: "numeric",
+};
 
 function properTimeFormatter(timeInMs, options) {
   const allSeconds = Math.round(timeInMs / 1000);
@@ -89,35 +156,18 @@ function properTimeFormatter(timeInMs, options) {
   return formattedTime;
 }
 
-function createUnixTimeID(unixTime) {
-  const unixDate = unixTime || Date.now();
-  const firstNumber = Math.round(unixDate / 1000);
-  const firstPart = firstNumber.toString().slice(3);
-  const randomNr = Math.floor(Math.random() * 1000);
-  const secondPart = randomNr.toLocaleString("en-US", {
-    minimumIntegerDigits: 3,
-    useGrouping: false,
-  });
-  const newId = Number(firstPart + secondPart);
-  return newId;
-}
-
-//needed?
-function disappearListContainer() {
-  const listContainer = document.getElementById("list-container");
-  listContainer.classList.add("list_container--passive");
-}
-
 export {
-  properTimeFormatter,
   createUnixTimeID,
   deleteEntry,
-  disappearListContainer,
-  dateOptions,
-  timeOptions,
+  deleteAllEntries,
+  loadBackup,
   generateRandomInteger,
   removeDuplicates,
   saveToLocalStorage,
   formReset,
+  handleEdit,
   timeReset,
+  properTimeFormatter,
+  dateOptions,
+  timeOptions,
 };
